refactor(group_store): extract namespace tests into namespace/tests.rs (#2480)
Phase 2/3 of #2480 (epic #2300). No behavior change — 48 test fns
and one namespace-only inline helper (`raw_namespace_dag_heads`)
move out of `group_store/tests.rs` into a focused
`group_store/namespace/tests.rs`, organized next to the code they
exercise.

## Outcomes

  before:  tests.rs               7,149 LOC, 150 test fns
  after:   tests.rs               4,766 LOC, 102 test fns
           namespace/tests.rs     2,406 LOC,  48 test fns

`cargo test -p calimero-context --lib` runs 315 tests, same as
phase 1. No tests added/removed; only relocated.

## What moved

Tests grouped by primary subject in the namespace cluster:

- DAG service: `namespace_dag_service_*` (5)
- Op log: `namespace_op_log_service_*` (3)
- Retry: `namespace_retry_service_*` (2)
- Nesting + topology: `namespace_nesting_*`, `is_descendant_of_*`,
  `reparent_group_*`, `collect_subtree_for_cascade_*`,
  `collect_visible_descendant_groups_*`,
  `create_recursive_invitations_*` (~15 total).
- Authorization: `authorized_for_state_op_*` (4),
  `is_authoritative_namespace_identity_*`.
- Recursive operations: `recursive_remove_*` (5).
- Governance ops: `governance_group_*` (4), `governance_apply_*`,
  `governance_rejects_*`, `execute_group_*` (2),
  `replica_applies_*`, `replica_op_log_*`, `replica_stale_head_*`.
- One-off: `min_acks_after_local_mutation_uses_publish_time_subscribers`.

The namespace-only helper `raw_namespace_dag_heads` moves alongside
(it's referenced only by the namespace_dag tests that came along).

## What stays in tests.rs

Anything whose primary subject is a different module: capabilities,
context tree/registration, deny list, group settings, group keys,
metadata, op-application pipeline, governance broadcast,
TEE policy lookup, signing keys, upgrades. These get extracted by
future #2300 sub-issues as those clusters consolidate.

## Path updates inside the moved tests

`super::local_state::*` → `super::super::local_state::*`
(`local_state.rs` is at group_store level, one further up from
namespace/tests.rs).

`namespace_governance::min_acks_after_local_mutation` (a stale
crate-internal path that resolved through the old mod tree)
→ `super::min_acks_after_local_mutation` (via the `pub(crate) use`
re-export added in phase 1).

## Lesson applied from #2306

The extraction script now explicitly handles the `#[cfg(test)]`
followed by `mod foo {` pattern — it skips the entire mod block
rather than attributing the attribute to the next top-level `fn`
it finds. The #2306 split script's lack of this check caused `mod
auto_follow_tests` (488 LOC) to be over-captured into the wrong
file; that bug cannot recur here. `mod auto_follow_tests` is
correctly left in `tests.rs`.

## Verification

- 315 passed; 0 failed; 0 ignored
- `cargo build --workspace --lib` green
- Zero unused-import warnings in the affected files
- `mod auto_follow_tests` correctly remained in `tests.rs` (script
  fix held)

Next: phase 3 is `cargo fmt` + final verification.
